+++ src/app/(business-dashboard)/settings/business-type/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import BusinessType from './business'
+
+export default function page() {
+    return (
+        <div className="flex flex-col items-center w-full px-8">
+            <BusinessType />
+        </div>
+    )
+}
