+++ src/app/(business-dashboard)/settings/device/page.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export default function page() {
+    return (
+        <div className="flex flex-col items-start w-full gap-8">
+            <div className="flex flex-col items-center gap-4 w-full">
+                <p className="text-[#0C394B] text-[16px] leading-[150%] font-[600]">
+                    Comming Soon!
+                </p>
+
+            </div>
+
+        </div>
+    )
+}
